fix loading components
classname should be added to components to give us the ability to use
them as styled component
fixed condition of logo animation cause when you don't set any else
statement, a false string would be set as the classname of the
component.
fixed color proptypes of circle wave loading,
added docz for components.